restore migrated tests without legacy tags
